Fixed some major bugs whereby I should have been iterating over an original graph, modifying a copy of it.  Beforehand I was iterating over the copy, which meant the stuff I was iterating over was changing as I was iterating over it.
In addition, I've added a 'calculate_node_stability' method which is
used in the graph comparison visualization to highlight the nodes that
are the 'volatile' nodes.  I've put in a first pass at this
visualization, but there are a couple of things to clean up.  In
addition, it is currently uni-directional in its calculation of WL
labels, whereas the actual kernel should be bidirectional.  I will fix
this shortly.